requirements.txt
opencv-python
ddgs
requests
Pillow
azure-storage-blob
python-dotenv
gradio
openai